refactor(ssid): extract phase2 site-variables cluster (phase 3/6)
Split the Phase 2 orchestrator (phase2_site_variables), its batched
write coordinator (_write_site_variables), and 8 pure helpers
(_compute_variable_plan, _extract_deviation_params, _build_skip_entry,
_get_cached_site_vars, _build_variable_entry, _display_variable_summary,
_print_conflicts, _group_entries_by_site) out of the parent module into
a new _SsidTemplatePhase2Cluster in src/ssid_consolidation/
_ssid_template_phase2.py. The cluster is registered in the parent's
self._clusters tuple so proxy access via __getattr__ preserves the
public API and existing test call sites.

_write_single_site_vars stays in the parent module because tests patch
mistapi through the parent's namespace (patch.object(_mod, "mistapi")),
which only intercepts functions whose __globals__ binding is the parent
module.

Compliance: D-/61 -> D/65 on the parent file. All 241 unit tests pass,
mypy --strict clean. F401 tolerances on backward-compat re-exports
match the phase 1/2 pattern.
